[store_adapter] refine some codes in store adapter
