+++ research/gaia/gremlin/gremlin_core/src/process/traversal/step/source.rs
@@ -19,10 +19,9 @@ use crate::process::traversal::step::util::StepSymbol;
 use crate::process::traversal::step::Step;
 use crate::process::traversal::traverser::{Requirement, Traverser};
 use crate::structure::codec::pb_chain_to_filter;
-use crate::structure::{Edge, Label, QueryParams, Vertex, ID};
+use crate::structure::{Edge, Label, LabelId, QueryParams, Vertex, ID};
 use crate::{FromPb, Partitioner};
 use bit_set::BitSet;
-use graph_store::common::LabelId;
 use pegasus::BuildJobError;
 use pegasus_common::downcast::*;
 use prost::alloc::str::FromStr;
